net: lib: lwm2m_client_utils: Move `lwm2m_update_signal_meas_objects`
Move `lwm2m_update_signal_meas_objects` to `ecid_signal_meas_info.c`
so that users don't have to enable the LTE event listener
(`CONFIG_LWM2M_LTE_EVENT_LISTENER`) in order to set the
signal measurement object.
