+++ subsys/net/lib/lwm2m_client_utils/lwm2m/lwm2m_lte_notification.c
@@ -12,7 +12,6 @@
 #include <net/lwm2m_client_utils.h>
 #include <lwm2m_engine.h>
 
-#include <modem/modem_info.h>
 #include <modem/lte_lc.h>
 
 #include <zephyr/logging/log.h>
@@ -24,81 +23,6 @@ static void lte_notify_handler(const struct lte_lc_evt *const evt);
 
 static K_SEM_DEFINE(rrc_idle, 0, 1);
 static bool measurement_scheduled;
-
-static void update_signal_meas_object(const struct lte_lc_ncell *const cell, uint16_t index)
-{
-	int obj_inst_id;
-	char path[sizeof("10256/65535/0")];
-
-	obj_inst_id = lwm2m_signal_meas_info_index_to_inst_id(index);
-
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/0", obj_inst_id);
-	lwm2m_engine_set_s32(path, cell->phys_cell_id);
-	/* We don't set the resource 1 as the lte_lc_ncell struct doesn't
-	 * contain MCC and MNC for calculating ECGI
-	 */
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/2", obj_inst_id);
-	lwm2m_engine_set_s32(path, cell->earfcn);
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/3", obj_inst_id);
-	lwm2m_engine_set_s32(path, RSRP_IDX_TO_DBM(cell->rsrp));
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/4", obj_inst_id);
-	lwm2m_engine_set_s32(path, RSRQ_IDX_TO_DB(cell->rsrq));
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/5", obj_inst_id);
-	lwm2m_engine_set_s32(path, cell->time_diff);
-}
-
-static void reset_signal_meas_object(uint16_t index)
-{
-	int obj_inst_id;
-	char path[sizeof("10256/65535/0")];
-
-	obj_inst_id = lwm2m_signal_meas_info_index_to_inst_id(index);
-
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/0", obj_inst_id);
-	lwm2m_engine_set_s32(path, 0);
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/2", obj_inst_id);
-	lwm2m_engine_set_s32(path, 0);
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/3", obj_inst_id);
-	lwm2m_engine_set_s32(path, 0);
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/4", obj_inst_id);
-	lwm2m_engine_set_s32(path, 0);
-	snprintk(path, sizeof(path), "10256/%" PRIu16 "/5", obj_inst_id);
-	lwm2m_engine_set_s32(path, 0);
-}
-
-int lwm2m_update_signal_meas_objects(const struct lte_lc_cells_info *const cells)
-{
-	static int neighbours;
-	int i = 0;
-
-	if (cells == NULL) {
-		LOG_ERR("Invalid pointer");
-		return -EINVAL;
-	}
-
-	if (cells->ncells_count == 0) {
-		LOG_DBG("No neighbouring cells found");
-		return -ENODATA;
-	}
-
-	LOG_INF("Updating information for %d neighbouring cells", cells->ncells_count);
-
-	for (i = 0; i < MAX_INSTANCE_COUNT && i < cells->ncells_count; i++) {
-		update_signal_meas_object(&cells->neighbor_cells[i], i);
-	}
-
-	/* If we have less neighbouring cells than last time, reset
-	 * object instances exceeding the current cell count
-	 */
-	if (cells->ncells_count < neighbours) {
-		for (i = cells->ncells_count; i < MAX_INSTANCE_COUNT && i < neighbours; i++) {
-			reset_signal_meas_object(i);
-		}
-	}
-	neighbours = cells->ncells_count;
-
-	return 0;
-}
 
 void lwm2m_ncell_schedule_measurement(void)
 {
